thryv/requirements.txt
Django==4.1
djangorestframework==3.13
google-cloud-texttospeech==3.9.1
google-auth==2.17.0
google-auth-oauthlib==0.5.2
google-auth-httplib2==0.1.0
google-api-python-client==2.94.0
psycopg2==2.9.3  # for PostgreSQL
django-cors-headers==3.10.0
